WIP: get `just build` to work
This is a work in progress to get `just build` to proceed. There are
still some glitches, so help wanted and feel free to push commits to
this branch.

Meanwhile, it might be handy to automate setup further with a tool like
[asdf](https://asdf-vm.com/). Here's an example project that does it:
https://github.com/birros/wazero-demo#setup

Once that's in, probably GitHub actions can test that a clean build
works on the intended platforms.

Signed-off-by: Adrian Cole <adrian@tetrate.io>